W8_A3/requirements.txt
google-generativeai
pdfplumber
python-docx